Add a `but watch` command with some debugging facilities to help with #8582.
Also: upgrade the `notify` crate to the latest version for good measure.
