+++ crates/but-cli/Cargo.toml
@@ -18,6 +18,7 @@ testing = ["dep:gitbutler-commit"]
 gitbutler-command-context.workspace = true
 gitbutler-stack.workspace = true
 gitbutler-project.workspace = true
+gitbutler-filemonitor.workspace = true
 but-settings.workspace = true
 but-core.workspace = true
 but-workspace.workspace = true
@@ -38,3 +39,4 @@ tracing.workspace = true
 dirs-next = "2.0.0"
 serde_json = "1.0.140"
 serde.workspace = true
+tokio.workspace = true
