+++ crates/but-settings/Cargo.toml
@@ -11,7 +11,7 @@ serde = { workspace = true, features = ["std"] }
 serde_json = { version = "1.0", features = ["std", "arbitrary_precision"] }
 serde_json_lenient = "0.2.3"
 gitbutler-fs.workspace = true
-notify = { version = "6.0.1" }
+notify = { version = "8.0.0" }
 tracing.workspace = true
 tokio = { workspace = true, features = ["macros", "rt"] }
